change fmod

+++ src/fmod/s21_fmod.h
@@ -11,5 +11,5 @@
 
 long double s21_fmod(double x, double y);
 long double s21_truncate(double x);
-long double s21_fma(long double x, long double y, long double z);
+long double s21_fma(double x, double y, double z);
 #endif
